croco: Coherently initialize objects ref counting to 1
Croco used to initialize most objects with a 0 ref-counting even though
it will behave exactly the same as an object with just one reference at
destruction time.
Exception was CRDocHandler which already followed this schema.

This is incoherent to what GLib does, given that at destruction time an
object with 1 reference will be still destroyed ad unref phase, and that
all the objects provide also a destroy function to ignore ref-counting
we can safely just ensure we always init objects with valid
ref-counting.

This required some minor code adaptation inside croco itself as we need
to ensure that we don't leak when an object is created and passed around
if referenced again.